+++ src/LIRSCache.java
@@ -70,7 +70,7 @@ public final class LIRSCache {
 	private static final boolean debug = false;
 
 	public LIRSCache(int inputCacheSize) {
-		int cacheSize = inputCacheSize * 1024 * 1024 / CacheSim.CACHE_BLOCK_SIZE;
+		int cacheSize = (int)(inputCacheSize * 1024L * 1024L / CacheSim.CACHE_BLOCK_SIZE);
 		this.maximumNonResidentSize = cacheSize*5;
 		this.maximumHotSize = (int)(cacheSize*0.9);
 		this.data = new HashMap<Long, Node>();
